move ThrottlerTransportAdapter classes
