+++ akka-cluster-tools/src/multi-jvm/scala/akka/cluster/singleton/ClusterSingletonManagerDownedSpec.scala
@@ -18,7 +18,7 @@ import akka.remote.testconductor.RoleName
 import akka.remote.testkit.MultiNodeConfig
 import akka.remote.testkit.MultiNodeSpec
 import akka.remote.testkit.STMultiNodeSpec
-import akka.remote.transport.ThrottlerTransportAdapter
+import akka.remote.testkit.Direction
 import akka.testkit._
 import akka.util.ccompat._
 
@@ -104,8 +104,8 @@ class ClusterSingletonManagerDownedSpec
 
     "stop instance when member is downed" in {
       runOn(first) {
-        testConductor.blackhole(first, third, ThrottlerTransportAdapter.Direction.Both).await
-        testConductor.blackhole(second, third, ThrottlerTransportAdapter.Direction.Both).await
+        testConductor.blackhole(first, third, Direction.Both).await
+        testConductor.blackhole(second, third, Direction.Both).await
 
         within(15.seconds) {
           awaitAssert {
@@ -116,7 +116,7 @@ class ClusterSingletonManagerDownedSpec
       enterBarrier("blackhole-1")
       runOn(first) {
         // another blackhole so that second can't mark gossip as seen and thereby deferring shutdown of first
-        testConductor.blackhole(first, second, ThrottlerTransportAdapter.Direction.Both).await
+        testConductor.blackhole(first, second, Direction.Both).await
         cluster.down(node(second).address)
         cluster.down(cluster.selfAddress)
         // singleton instance stopped, before failure detection of first-second
