+++ akka-cluster/src/multi-jvm/scala/akka/cluster/RestartNode3Spec.scala
@@ -18,7 +18,7 @@ import akka.actor.RootActorPath
 import akka.cluster.MemberStatus._
 import akka.remote.testkit.MultiNodeConfig
 import akka.remote.testkit.MultiNodeSpec
-import akka.remote.transport.ThrottlerTransportAdapter.Direction
+import akka.remote.testkit.Direction
 import akka.testkit._
 import akka.util.ccompat._
