+++ akka-cluster/src/multi-jvm/scala/akka/cluster/SplitBrainSpec.scala
@@ -10,7 +10,7 @@ import com.typesafe.config.ConfigFactory
 import language.postfixOps
 
 import akka.remote.testkit.MultiNodeConfig
-import akka.remote.transport.ThrottlerTransportAdapter.Direction
+import akka.remote.testkit.Direction
 import akka.testkit._
 
 final case class SplitBrainMultiNodeConfig(failureDetectorPuppet: Boolean) extends MultiNodeConfig {
