+++ akka-cluster/src/multi-jvm/scala/akka/cluster/StreamRefSpec.scala
@@ -14,7 +14,7 @@ import com.typesafe.config.ConfigFactory
 import akka.Done
 import akka.actor.{ Actor, ActorIdentity, ActorLogging, ActorRef, Identify, Props }
 import akka.remote.testkit.MultiNodeConfig
-import akka.remote.transport.ThrottlerTransportAdapter.Direction
+import akka.remote.testkit.Direction
 import akka.serialization.jackson.CborSerializable
 import akka.stream.Materializer
 import akka.stream.RemoteStreamRefActorTerminatedException
