+++ akka-remote-tests/src/multi-jvm/scala/akka/remote/AttemptSysMsgRedeliverySpec.scala
@@ -15,7 +15,7 @@ import akka.actor.ActorRef
 import akka.actor.Identify
 import akka.actor.PoisonPill
 import akka.actor.Props
-import akka.remote.transport.ThrottlerTransportAdapter.Direction
+import akka.remote.testkit.Direction
 import akka.testkit._
 
 object AttemptSysMsgRedeliveryMultiJvmSpec extends MultiNodeConfig {
